+++ GrowDiary.Web/Program.cs
@@ -100,6 +100,8 @@ builder.Services.AddSingleton<NotificationSettingsRepository>();
 builder.Services.AddSingleton<AppSettingsRepository>();
 builder.Services.AddSingleton<WaterProfileStore>();
 builder.Services.AddScoped<GrowCostService>();
+builder.Services.AddScoped<MischplanService>();
+builder.Services.AddScoped<SopDueService>();
 builder.Services.AddSingleton<SolutionStabilityAnalyzer>();
 builder.Services.AddScoped<TrendWatchRunner>();
 builder.Services.AddSingleton<TentSensorHardwareSyncService>();